Add the concept of soft and hard constraints.
Soft constraints may be violated, hard canstratints may not. Soft
constraint violations are displayed in yellow and hard in red.  In the
future when the schedule can be uploaded via the web, hard constraint
violations will cause the upload to fail, soft constraint violations
will only display as warnings and perhaps a confirmation page.