Fix error in dumping table type information
No table type information was dumped for the early regional tournaments,
so imports don't work right. It turns out that when getting the names of
the tables from metadata.getTables() the table names come back
lowercase, but metadata.getColumns() wants the table name in all caps.
Modified the dump routine to try as is, upper case and then lower case
if no columns are found.