Remove null rows from subjective tables
When data is uploaded and all scores for a team have been deleted and
the team is not a No Show, then the row shouldn't be in the database and
causes some special cases in the code. These rows are now removed after
data is entered upon upload.

This fixes ticket:13