Prevent entering scores by teams that don't advance in playoffs.
Addresses trac ticket #77.
Adds query to check if a team is listed in the playoffdata table for a
particular round. If they are not listed, they did not advance (according to
currently-entered scores) and therefore you should not be able to enter a score
for them. Update to scoreEntry.jsp uses the new query to make that check and
throw an exception with an appropriate message.