Fix some issues with redirections to setup and the tests
Once I got selenium working again at home I found a couple of corner
cases where we'd get a redirect loop. I also found that we really need
to login after initializing the database and always do it after
initialization, so that the database exists and we're always logged in.