Fix thread safety and some cleanup
Access to the list of IPs AND the expiration time needs to be inside the
lock. Otherwise you can get a race reading or updating the expiration
time.

Handle the IOException directly inside getAllIPs(). Callers can't really
do much with this exception anyway, so just return an empty list.

Make sure to return an unmodifiable collection so that callers can't
modify the cache of IPs and break our thread safety.