Rework checking seeding runs and namespace the session variables
Turns out it doesn't work very well to check seeding rounds for a newly
created division as the division needs to be initialized first at which
point it doesn't matter if you've checked or not. So don't allow the
user to check.

Start moving session data into a custom class to make it easy to clean
up and not conflict with other variables.