Handle odd teams and performance rounds
If there is an odd number of teams and an odd number of performance
rounds, then 1 team needs to run without an opponent. So allow this once
and see if we can come up with a schedule. This also checks that a team
can stay around to compete against the team if needed.