Resolve variable and goal references late
The variable and goal references can't be resolved right away at
construction time as the referenced objects may not exist yet. So now
the lookup is done when the object is requested. One might try and cache
the result of the lookup at some point, but then we need to make sure we
do locking for thread safety.