Do insert and update of performance scores in a transaction
This should remove our problems with multiple updates and inserts
happening at the same time by grabbing a serializable lock on the
database for the row that we're going to change and then updating first
and if that doesn't change anything do the insert. The update will lock
the row that we want to change so that other transactions working with
the same row will have to wait.