Make it clear when specifying judging station vs. event division
When a team is moved to a tournament that uses event divisions the user
is prompted for both judging station and event division. In many cases
these are the same so it appears that the pages are asking for the same
thing. I bolded the text to make it clear.