Pass on the 3rd 4th place parameter when creating a division
Turns out the form parameter wasn't passed on. Now the form parameter is
stored in the session in InitializeBrackets when creating a new playoff
division and then cleared when the division is actually created. The
session variable is also cleared in the index. This should keep it from
leaking.

task #23: 3rd & 4th place bracket didn't work at state 
http://github.com/jpschewe/fll-sw/issues/issue/23