Fix an issue where table labels were always being placed *above* their team's leaf
The issue was that the BracketQuery servlet was now returning json that had "dbLine" in all lower case. Hence, when the javascript read the "dbLine" attribute, it was undefined, which we then passed to placeTableLabel as the dbLine value. undefined % 2 is NaN, and since NaN is not equal to 1, we fall into the else case which assumes that the value was even.