Uploading teams only adds teams
Instead of deleting all team information, the team upload now adds the
teams to the existing database without changing any existing teams. This
checks that no duplicate teams will be added first and notifies the user
of the duplicate teams so that they can be corrected in the spreadsheet.


Fixes  #230: Allow one to upload teams without wiping out the existing
teams [sf#195] 
http://github.com/jpschewe/fll-sw/issues/issue/230