Remove duplicate entries from schedule when loading from the database
When loading from the database there was an extra call to add team
schedule information to the underlying list. This caused each entry to
show up twice in the schedules.

Added checking to the add to watch for duplicates and ignore them with a
warning log message.

task #14: Add the ability to print the schedules 
http://github.com/jpschewe/fll-sw/issues/issue/14