Check if summary data is out of date and prompt the user
The user can then decide if the summarized scores will be computed or
not. To avoid infinite loops with multiple changes to the database, once
the user is prompted we don't check if scores need to be summarized
again until they start back at the index page.

task #351: Detect when summary data is out of date 
http://github.com/jpschewe/fll-sw/issues/issue/351