Properly count the number of teams that a judge has seen
When counting the number of teams one needs to look at which teams the
judge was supposed to see and only count them. This entails checking
against the team numbers in the judging station in case a judge is
assigned 2 judging groups in the same category and only scores 1 judging
group.

Fixes #367: Problem with 1 judge and 2 judging groups 
http://github.com/jpschewe/fll-sw/issues/issue/367