Construct date objects when functions are required
Because the schedule is serialized to JSON and then deserialized the
objects become just data. This is OK for most things, except when one
needs to call functions on the objects like I do for dates. So new date
objects are created whenever the functions are needed and we may have
previously serialized to JSON.

task #22: Be able to override auto generated finalist schedule 
http://github.com/jpschewe/fll-sw/issues/issue/22