Allow one to download the subjective data without all judges being assigned
When some judges are assigned by the subjective web application we will
have cases where there are not enough judges assigned for all teams, but
still want to enter some scores with the subjective desktop application.
I removed the check for judges being assigned so that we can do this.

Fixes #406