Load room and public/private category information
When loading the finalist schedule information, if a schedule exists,
load the room information and which categories are public and which are
private.

This starts the work for task #447 to be completed at a later date.
These changes work and don't get in the way, so I've left them in.

task #22: Be able to override auto generated finalist schedule 
http://github.com/jpschewe/fll-sw/issues/issue/22