Make sure that applications exit in a way that they can be used elsewhere
The Launcher will want to call both SchedulerUI and SubjectiveFrame, but
not exit the JVM when they close. So I needed to modify how I close
these windows to not call exit, but rather to call setVisible(false).
Then I needed to use both a ComponentListener and a WindowListener to be
sure that I get the events on all operating systems.

task #467: Single application to launch everything 
http://github.com/jpschewe/fll-sw/issues/issue/467