Note that only the JRE is required
Tomcat now (perhaps for some time) contains a compiler, so the JDK is
not required to run Tomcat. Therefore our requirement for the JDK has
been reduced to the JRE. Furthermore with the inclusion of the Launcher
application and it's ability to set JAVA_HOME for launching Tomcat there
is no need for the user to set JAVA_HOME themselves as the launcher will
just figure it out.


task #466: Create windows executables 
http://github.com/jpschewe/fll-sw/issues/issue/466