Rename "Event Division" to "Award Group" and "Playoff Division" to "Playoff Bracket"
I believe I've made all of the changes now. Some related changes willb e
made when task #486 is completed.

This came out of the meeting with Vicki and Jeannie to make the terms
less confusing. The word "division" has been over used and some people
think of this concept to mean grouping by age or experience. Instead
stick with "group". The playoff should make it clear that a bracket is
the whole head to head competition down to 1 winner. No need to have
another "division" name. 2 teams competing on a table is a "match".

Fixes #496

Fixes #497