Reword judging group and award group assignment info
On the Admin Index there are links to edit the award groups and judging
groups. The wording suggested that teams may not be assigned to an award
group or a judging group. This isn't possible, teams are always in an
award group and a judging group for the current tournament. So the links
to these pages have been updated to be clear that they allow one to
change the assignments.

Fixes #515