Rework how score sheets are generated
Remove some extra arrays that aren't needed. In most cases we can only
create a single object or use a method call with the PdfPTable that is
being created.

Also put all of the goals in a single table when specifying the
challenge description. Then the code that adds the goals to the score
sheet can just add this table without needing to know anything about how
the table is laid out.