Allow the performance duration to be a list
Normally this is just set to a single number, but it can be set to the
list {7, 8} to allow the schedule to have 2 teams every 15 minutes. This
allows some more time for teams without having a lot of odd times as
about half of the times should be on the quarter hour.