Rework remote control page
The remote control page has been reworked to support multiple head to
head brackets on the same display. In the process of doing this I
reworked the data that is stored in the application to contain all
information for a display in a single object. The display pages now need
to be reworked to use this new object.