Create copy of web dir for jasper task
For some reason the jasper task is hanging onto the libraries used in
the comiplation of the JSP files after it's finished. This causes the
dist target to fail on windows because the files can't be deleted. So
instead I'm copying everything to another directory at the top and then
cleaning that up on exit.