Refactor how BracketInfo is created
Instead of having to construct a BracketInfo object in javascript I made
BracketInfo a parent class to BracketData and then just pass a JSON
version of BracketData to the javascript which can then be passed back
to the H2HUpdateWebSocket and there's no concern about the data
structure becoming inconsistent.

I also reworked the remote control brackets to move it's Java code out
of the JSP.

Issue #552, #536.