Fix the public HTML finalist schedule
It was setup to only work as a display page. We're not using it as a
display page anymore, but it can be useful for a link on the public
index and the reporting page. So I reworked it to get the division
information from a parameter so that the page properly displays.

Fixes #554