Rework message heirarchy to match specification
Got some information from @poelstra about the message specification.
Because of this I changed class names to match his message names.

To handle the deserialization I needed to upgrade to jackson 2.6+, so I
went all the way to the latest release.

Publish ack messages are recognized, but nothing is done with them.

Issue #570