Change how missing opponents in performance are reported
After more thought about this, I think that trying to determine when to
have a team stay around is more than the software should be doing. All
the software should do is to let the user know that a team doesn't have
an opponent. It's up to the user to decide what to do.

Fixes #534