Allow the user to specify the filename suffix on subjective score sheets
This is useful when the user wants to put paper colors on the subjective
sheets. The default mapping is what Minnesota uses and depends on the
names of the subjective categories in the challenge description to not
change. If the subjective category name is not recognized, then the
default is no suffix. However the user can type in whatever they want.

Fixes #688