Handle quotes in team names for score verification
Team names with double quotes in them are a problem for the unverified
runs javascript because the team name is put into a javascript string.
If the double quotes aren't properly escaped, then the double quotes in
the team name end the string and break the code.

The solution here was to convert the JSP to a Servlet because it's
mostly Java anyway. Then use StringUtils to escape the string per the
javascript rules.

Fixes #711