Lock rubric range titles in the header
Create a separate table in the header for the rubric range titles. Then
repeat them in the first row of the scoring table with the row set to
hidden via CSS. This still keeps everything aligned because we are using
"table-layout: fixed" which uses the first row of data to determine
table column widths.

Fixes #746