Use forward instead of redirect
This allows the form parameter to be automatically passed along and
avoid SpotBugs warnings about passing request headers.

This does have the downside of the location in the browser address bar
not showing the jsp page. However it should be faster and allows for
more parameter passing internal to the container.

Issue #313