Redirect to the index when summarizing scores from the index
Ensure that when score summarization is initiated from the index.jsp,
redirect back to index.jsp. This is done by checking the referrer and
clearing the redirect session variable if the referrer ends with
"index.jsp". This handles the case where the last score summarization
was initiated from a report creation and the session variable is still
around because the index page hasn't been reloaded.

Issue #767