Always set the award groups
The award group for teams should always be set when loading a schedule.
The only case where this shouldn't happen is at state and we can fix
that by adding the award group column to the schedule or by setting them
all after the fact, either of which is what we normally need to do.

Issue #775