Don't add 's' to the end of the next tournament level
Take exactly what the user wants the next level text to be and put that
in the awards report. So when one wants to see "Sectionals" in the
advancing teams section of the awards report, the "next level" field in
the edit tournaments page needs to have "Sectionals" rather than
"Sectional" in it.

Fixes #812