Remove application cache from finalist scheduling workflow
The application cache is now only supported on secure sites. We aren't
running a secure site for a number of reasons and so this support is
being removed. If we want to run this workflow offline in the future we
should turn it into a single page web application like the subjective
scoring application.

Issue #826