Rework non-numeric nominees to not use an iframe
Using an iframe causes 2 separate sets of javascript objects to exist
and they don't cross the iframe boundary. Thinking about this it makes
perfect sense. So I've just copied the HTML into the non-numeric
nominees page. There isn't a lot to copy, so no big deal. This does
remove my funny logic for removing the navbar as well and makes it
easier to customize this page in the future.

Issue #810