Remove "name" from non-numeric category
Most everything is already done by title anyway, so the name is not
really a useful attribute and just adds more information the user needs
to put in the challenge description.

Eventually the javascript code should have "name" replaced with "title"
to make it clear there as well. I didn't do that now as refactoring the
javascript is much harder than refactoring the Java.

Issue #846