Filter all teams on scoreboard by judging group
Previously the scoreboard was filtered by award group. However filtering
by judging group is more useful as the judging groups are equal to the
award groups at regional tournaments, but at state the judging groups
are separate. This allows us to shorten the all teams list at state and
at the regional tournaments.

If the judging group is different from the award group it is now
displayed in parenthesis on the all teams page.

Issue #820