Move where bracket output is generated
There are some methods on BracketData that need to be called after
construction, but before the output is generated. So I added a generate
method that is to be called when the output should be generated and then
the output can be later retrieved when the connection is no longer
available.

Issue #737