Fix issue with double question marks in comments
When jQuery sees 2 question marks in a row in some JSON it assumes that
the data is actually JSONP and interprets the two question marks as a
placeholder for a callback. This breaks the the parser. The fix is to
specify the content type and then jQuery doesn't need to guess.

Issue #897