Use grid layout for subjective application.
This removes the need for absolute positioning and javascript to pad
the sections. The key to making this work is that all parent elements
of the main grid container need to have their height set to 100% and
the container needs to have it's height set to 100% instead of
100vh. Using 100vh on a desktop browser works, but not on a mobile
browser. I found this solution at
https://stackoverflow.com/a/68749468/2965897