Support empty awards script sections
Don't treat the empty string as NULL when storing in the database. The
empty string is a valid value and will result in the text being
empty. If the text value isn't specified, then it's deleted from the
database, so there is a clear difference between not specified and
specified as an empty string.

Issue #1020