Fix issue with overwriting text field values
Need to use backspace to clear out the existing data and then write
the new value. Can't use the clear method as focus is lost after the
call and the javascript event handler runs that replaces the value.
Using ctrl-a to select and then write doesn't appear to be working
either.

Issue #1074