Sort out issues between UUIDs used for display and head to head
Make it clear that UUIDs for head to head may match UUIDs for the
display, but the inverse is not necessarily true.

Rework code to make it clear when a head to head web socket is tied to
a display and is getting information on what brackets to display from
the remote control page.

Issue #1091