Refactor max run number code
Find all places that I look for the max run number in a tournament and
use the same method call.

Rename the methods to make it clear which is for a tournament and
which is for a team. This caused me to find some bugs where the wrong
one was used.

Factor out the setting of a common page variable into a class and it's
own method to start centralizing some duplicate setting of variables.

Issue #1229